==> 1-crewaling-product-title-fa.py.ipynb <==
# import libs
from urllib.request import urlopen
from bs4 import BeautifulSoup
from urllib.error import HTTPError
from urllib.error import URLError
import urllib.parse
import re
import http.client as http
http.HTTPConnection._http_vsn = 10
http.HTTPConnection._http_vsn_str = 'HTTP/1.0'

import ssl
ssl._create_default_https_context = ssl._create_unverified_context

import time

# crowl site for titles
pages = set()
titlesave=set()
stop = False

def getlinks(pageurl):
    global pages
    global site
    global stop

    try:

        html = urlopen(pageurl)
        site = pageurl
    except HTTPError as e:
        if e.code == 404:
            print("صفحه پیدا نشد.")
        elif e.code == 308:
            redirect_url = e.headers.get('Location')
            if redirect_url:
                print(f"Redirecting to: {site + redirect_url}")
                getlinks(site + redirect_url)
            else:
                print("Redirect URL not found.")
        else:
            print(e)
    except URLError as e:
        print("The server could not be found!")
    except AttributeError as e:
        print("Tag was not found.")
    else:
        bs = BeautifulSoup(html.read(), 'html.parser')
        links = bs.find_all('a', href = re.compile(r'^\/search\/categor(ies|y).*'))

        for link in links:
            if 'href' in link.attrs:
                if link.attrs['href'] not in pages:
                    if stop == True:
                        break
                    newpage = link.attrs['href']
                    print(newpage)
                    pages.add(newpage)
                    newurl = urllib.parse.quote(newpage)
                    # level 2

                    try:
                        html2 = urlopen(site + newurl)
                    except HTTPError as e:
                        if e.code == 404:
                            print("صفحه پیدا نشد.")
                        elif e.code == 308:
                            redirect_url = e.headers.get('Location')
                            if redirect_url:
                                print(f"Redirecting to: {site + redirect_url}")
                                getlinks(site + redirect_url)
                            else:
                                print("Redirect URL not found.")
                        else:
                            print(e)
                    except URLError as e:
                        print("The server could not be found!")
                    except AttributeError as e:
                        print("Tag was not found.")
                    else:
                        bs2 = BeautifulSoup(html2.read(), 'html.parser')
                        links2 = bs2.find_all('a', href = re.compile(r'.*(\/product(s)?\/).*', re.IGNORECASE))
                        for link2 in links2:
                            if 'href' in link2.attrs:
                                if link2.attrs['href'] not in pages:
                                    newpage2 = link2.attrs['href']
                                    pages.add(newpage2)
                                    newurl2 = urllib.parse.quote(newpage2)
                                    # level 3

                                    try:
                                        html3 = urlopen(site + newurl2)
                                    except HTTPError as e:
                                        if e.code == 404:
                                            print("صفحه پیدا نشد.")
                                        elif e.code == 308:
                                            redirect_url = e.headers.get('Location')
                                            if redirect_url:
                                                print(f"Redirecting to: {site + redirect_url}")
                                                getlinks(site + redirect_url)
                                            else:
                                                print("Redirect URL not found.")
                                        else:
                                            print(e)
                                    except URLError as e:
                                        print("The server could not be found!")
                                    except AttributeError as e:
                                        print("Tag was not found.")
                                    else:
                                        bs3 = BeautifulSoup(html3.read(), 'html.parser')
                                        title3 = bs3.find('h1')
                                        print(title3.get_text())
                                        titlesave.add(title3.get_text())
                                        if len(titlesave) >= 1000:
                                            stop = True
                                            break

                                        print("_" * 50)
                                        time.sleep(2)  # Add a delay of 1 second between requests

getlinks('https://timcheh.com')


import random

# Generating a set with 20 random values
random_values = random.sample(titlesave, 20)
print(random_values)

random_values2 = random.sample(titlesave, 20)
print(random_values2)

import pandas as pd

# Convert the set or list to a pandas DataFrame
df = pd.DataFrame(random_values, columns=['Values'])

# Save the DataFrame to an Excel file
df.to_excel('output.xlsx', index=False)

# Convert the set or list to a pandas DataFrame
df = pd.DataFrame(random_values2, columns=['Values'])

# Save the DataFrame to an Excel file
df.to_excel('output2.xlsx', index=False)
