==> 4_compare_libs.ipynb <==
# import libs
import pandas as pd
import pprint

from google.colab import drive
drive.mount('/content/drive')

# get titles from excels
def get_records_from_excel(file_path):
    df = pd.read_excel(file_path)
    values_set = list(df.iloc[:, 0])
    return values_set


records = get_records_from_excel("/content/drive/My Drive/project/finals/output.xlsx")
print(records)
records2 = get_records_from_excel("/content/drive/My Drive/project/finals/output2.xlsx")
print(records2)

!pip install hazm

# import stopwords
import re
from nltk.corpus import stopwords
import nltk
nltk.download('stopwords')
from hazm import stopwords_list

# خواندن کلمات فارسی از فایل persion.txt
# with open('/content/drive/My Drive/project/finals/persian.txt', 'r', encoding='utf-8') as file:
#     persian_stopwords = file.read().split()

# لیست کلمات استاپ ورد انگلیسی از NLTK
english_stopwords = set(stopwords.words('english'))


from hazm import word_tokenize
from hazm import Normalizer
from hazm import Lemmatizer

# Function to process sentences
def process_sentences(records):
    normalizer = Normalizer()
    lemmatizer = Lemmatizer()
    processed_sentences = []

    for sentence in records:
        normalized_sentence = normalizer.normalize(sentence)
        tokens = word_tokenize(normalized_sentence)
        processed_tokens = []

        for word in tokens:
            if re.search(r'[\u0600-\u06FF]', word):  # Persian
                word = lemmatizer.lemmatize(word)
                if word not in stopwords_list():
                    processed_tokens.append(word)
            elif word.lower() not in english_stopwords:  # English
                processed_tokens.append(word)

        processed_sentences.append(" ".join(processed_tokens))

    return processed_sentences

# Process sentences
dic1 = process_sentences(records)
dic2 = process_sentences(records2)

print(dic1)
print('\n\n')
print(dic2)

!pip install fuzzywuzzy

from fuzzywuzzy import process

# Find the most similar record and store results in a DataFrame
results = []

for str1 in dic1:
    highest = process.extractOne(str1, dic2)
    results.append([str1, highest[0], highest[1]])

# Create a DataFrame
df_results = pd.DataFrame(results, columns=['Record from Table 1', 'Most Similar Record from Table 2', 'Similarity'])

# Display the DataFrame
df_results.head()

import difflib

# Find the most similar record using difflib and store results in a DataFrame
results = []

for str1 in dic1:
    matches = difflib.get_close_matches(str1, dic2, n=1, cutoff=0)
    if matches:
        most_similar = matches[0]
        similarity = difflib.SequenceMatcher(None, str1, most_similar).ratio()
        results.append([str1, most_similar, similarity])
    else:
        results.append([str1, None, 0])

# Create a DataFrame
df_results = pd.DataFrame(results, columns=['Record from Table 1', 'Most Similar Record from Table 2', 'Similarity'])

# Display the DataFrame
df_results.head()

from nltk.metrics.distance import edit_distance

# Function to calculate Levenshtein distance
def calculate_levenshtein_distance(text1, text2):
    levenshtein_distance = edit_distance(text1, text2)
    return levenshtein_distance

# Find the most similar record using Levenshtein distance and store results in a DataFrame
results = []

for str1 in dic1:
    min_distance = float('inf')
    most_similar = None
    for str2 in dic2:
        distance = calculate_levenshtein_distance(str1, str2)
        if distance < min_distance:
            min_distance = distance
            most_similar = str2
    similarity = 1 - min_distance / max(len(str1), len(most_similar))  # Similarity as normalized value
    results.append([str1, most_similar, similarity])

# Create a DataFrame
df_results = pd.DataFrame(results, columns=['Record from Table 1', 'Most Similar Record from Table 2', 'Similarity'])

# Display the DataFrame
df_results.head()
